requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==3.2.13
djangorestframework==3.13.1
importlib-metadata==4.11.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.2.0
zipp==3.8.0
psycopg2-binary==2.9.3
pytz==2022.1